requirements.txt
durable_rules
flask
request